Add neurosift clear cache test and ts declarations

+++ src/electron/frontend/core/components/Neurosift.js
@@ -80,7 +80,9 @@ export class Neurosift extends LitElement {
     render() {
         // Clear neurosift cross-session storage database
         // see https://github.com/NeurodataWithoutBorders/nwb-guide/issues/974
-        indexedDB.deleteDatabase("neurosift-hdf5-cache");
+        if (this.url) {
+            indexedDB.deleteDatabase("neurosift-hdf5-cache");
+        }
 
         return this.url
             ? html` <div class="loader-container">
